Add traverse2With
Add `traverse2With`, which is reminiscent of
`Control.Lens.Traversal.traverseOf`, but it upgrades a polymorphic
`Applicative` traversal (essentially an unwrapped `Bazaar (->)`) to a
`Biapplicative` traversal.

Why do we put this here? It seems like the sort of thing
that would be more at home in `lens`, with some profunctory
generalization. But we're using some fairly heavy and unpleasant
machinery to upgrade a function with a `traverse`-like type to
one with a `traverse2`-like type; I figure we might as well let
users have the more general version. If we end up using something
like this in `lens`, we'll probably need to build the machine again
to add the profunctor business, but I guess that's life.