Define missing Functor/Foldable/Traversable instances in the test suite
A future version of GHC will add a quantified superclass to `Bifunctor`, which
causes the `bifunctors` test suite to no longer typecheck due to missing
`Functor` instances. This patch adds them, along with `Foldable` and
`Traversable` instances to prepare for a future in which `Bifoldable` and
`Bitraversable` have quantified superclasses as well.